refactor create pending order
