+++ atdd-by-example/order/src/main/java/com/atdd/order/application/CreateOrderServiceImpl.java
@@ -1,0 +1,17 @@
+package com.atdd.order.application;
+
+import com.atdd.order.domain.PendingOrder;
+import com.atdd.order.domain.PendingOrderRepository;
+import com.atdd.order.infra.PendingOrderRepositoryMemoryImpl;
+
+class CreateOrderServiceImpl implements CreateOrderService {
+
+    private PendingOrderRepository pendingOrderRepository = new PendingOrderRepositoryMemoryImpl();
+
+    @Override
+    public PendingOrder createPendingOrder(PendingOrderRequest request) {
+        PendingOrder pendingOrder = new PendingOrder(request.getProductId(), request.getQuantity());
+        return pendingOrderRepository.save(pendingOrder);
+    }
+
+}
