+++ atdd-by-example/order/src/test/java/com/atdd/order/application/createOrderServiceTest.java
@@ -1,10 +1,7 @@
 package com.atdd.order.application;
 
 import com.atdd.order.domain.PendingOrder;
-import com.atdd.order.domain.PendingOrderRepository;
 import org.junit.jupiter.api.Test;
-
-import java.util.concurrent.atomic.AtomicLong;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -22,29 +19,4 @@ public class createOrderServiceTest {
         assertThat(pendingOrder.getId()).isPositive();
     }
 
-    private class CreateOrderServiceImpl implements CreateOrderService {
-
-        private PendingOrderRepository pendingOrderRepository = new PendingOrderRepositoryMemoryImpl();
-
-        @Override
-        public PendingOrder createPendingOrder(PendingOrderRequest request) {
-            PendingOrder pendingOrder = new PendingOrder(request.getProductId(), request.getQuantity());
-            return pendingOrderRepository.save(pendingOrder);
-        }
-
-    }
-    private static class PendingOrderRepositoryMemoryImpl implements PendingOrderRepository {
-
-        AtomicLong atomicId = new AtomicLong(1);
-
-        @Override
-        public PendingOrder save(PendingOrder pendingOrder) {
-            pendingOrder.assignId(nextId());
-            return pendingOrder;
-        }
-
-        private long nextId() {
-            return atomicId.getAndIncrement();
-        }
-    }
 }
